+++ packages/workspace-fusion-modules/src/lib/modules/fusionBookmarks/BookmarkModule.tsx
@@ -2,12 +2,14 @@ import { FusionBookmark, FusionMediator, WorkspaceViewController } from '@equino
 import { HeaderContext, useWorkspaceHeaderComponents } from '@equinor/workspace-fusion';
 import { QueryClient, QueryClientProvider } from 'react-query';
 
-import { BookmarksHook, HeaderIcon } from './components';
+import { BookmarkLoader, HeaderIcon } from './components';
 
 export type BookmarksModuleConfig = {
 	searchParam?: string;
 	getBookmark: (id: string, signal?: AbortSignal) => Promise<FusionBookmark<any>>;
 };
+
+const moduleName = 'bookmarks';
 
 /**
  * Will load bookmarks from url
@@ -16,7 +18,7 @@ export function BookmarksModule(config: BookmarksModuleConfig) {
 	const queryClient = new QueryClient();
 
 	return {
-		name: 'bookmarks',
+		name: moduleName,
 		setup: (
 			mediator: FusionMediator<any, any, any>,
 			appKey: string,
@@ -30,15 +32,23 @@ export function BookmarksModule(config: BookmarksModuleConfig) {
 					return (
 						<QueryClientProvider client={queryClient}>
 							{/* Injects icon into workspace header */}
-							<HeaderContext.Provider value={{ ...context, BookmarksIcon: HeaderIcon }}>
-								{/* Halts the component while applying a bookmark from url */}
-								<BookmarksHook
+							<HeaderContext.Provider
+								value={{
+									...context,
+									icons: [
+										...context.icons.filter((s) => s.name !== moduleName),
+										{ name: moduleName, Icon: HeaderIcon, placement: 'right' },
+									],
+								}}
+							>
+								{/* Halts the component while applying bookmark from url */}
+								<BookmarkLoader
 									getBookmark={config.getBookmark}
 									searchParam={config?.searchParam ?? 'bookmarkId'}
 									mediator={mediator}
 								>
 									{children}
-								</BookmarksHook>
+								</BookmarkLoader>
 							</HeaderContext.Provider>
 						</QueryClientProvider>
 					);
