+++ packages/workspace-fusion-modules/src/lib/modules/fusionBookmarks/index.ts
@@ -1,1 +1,1 @@
-export * from './module';
+export * from './BookmarkModule';
