+++ packages/workspace-fusion-modules/src/lib/modules/index.ts
@@ -1,1 +1,1 @@
-export { BookmarksModule } from './fusionBookmarks/module';
+export { BookmarksModule } from './fusionBookmarks/BookmarkModule';
